Moving managing data for Navigation entities into hook
